Move/rename all feature extractor related modules.
Now they're all in the same subdirectory instead of just the EfficientNet-detail modules being separated.

All extractor classes are now importable from spacer.extractors, which should remain the case even as more refactors are done in subsequent commits.
