+++ spacer/extractors/utils.py
